chore: rename to index_commits
