+++ ee/tabby-webserver/src/service/background_job/index_commits.rs
