+++ ee/tabby-webserver/src/service/background_job/mod.rs
@@ -1,9 +1,9 @@
-mod commit_index;
 mod daily;
 mod db;
 mod git;
 mod helper;
 mod hourly;
+mod index_commits;
 mod index_garbage_collection;
 mod license_check;
 mod third_party_integration;
